backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
httpx
pytest
python-multipart
python-dotenv
sqlmodel
psycopg2-binary
supabase
